backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.14.0
greenlet>=3.1.0

# Auth
python-jose[cryptography]>=3.3.0
msal>=1.28.0
pyotp>=2.9.0
qrcode[pil]>=7.4

# HTTP
httpx>=0.28.0

# Email
aiosmtplib>=3.0.0

# Redis / Task queue
redis>=5.2.0
celery>=5.4.0

# AI / LLM
langgraph>=0.2.0
langchain-core>=0.3.0
litellm>=1.55.0
# Added by A3 (LangGraph checkpointing): AsyncPostgresSaver + psycopg3 connection pool
langgraph-checkpoint-postgres>=2.0.0

# WebSocket
websockets>=14.0

# Logging
structlog>=24.0.0

# Phase 5 / W5.2 — distributed tracing
# Production deployments configure exporters via standard OTel env vars
# (OTEL_EXPORTER_OTLP_ENDPOINT, OTEL_SERVICE_NAME, etc.) and set
# ARCHON_OTEL_EXPORTER=otlp. tracing.py degrades to no-op without these.
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-instrumentation-fastapi>=0.51b0

# Scheduling
croniter>=2.0.0

# Misc
python-multipart>=0.0.17

# Cloud / Storage connectors
boto3>=1.35.0

# Vault
hvac>=2.3.0

# PII detection / redaction
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Security pins (Snyk)
h11>=0.16.0          # Critical: HTTP Request Smuggling (CVE in h11<0.16.0)
idna>=3.7            # Medium: Resource Exhaustion
jinja2>=3.1.6        # Medium: Template Injection (3 CVEs in jinja2<3.1.6)
orjson>=3.11.6       # Medium: Uncontrolled Recursion
pyasn1>=0.6.2        # High: Resource Allocation
setuptools>=78.1.1   # High: Code Injection; Medium: ReDoS, Directory Traversal
tqdm>=4.66.3         # Medium: Injection
urllib3>=2.6.3       # High(x3) + Medium(x2): multiple CVEs in urllib3<2.6.3
# NOTE: ecdsa@0.19.1 (High: timing attack, missing encryption) is pulled in as a
# transitive dep of python-jose, but is NOT used at runtime because
# python-jose[cryptography] is specified above — the cryptography backend takes
# precedence, making the ecdsa code paths unreachable.

# A15: enable safe expression eval in node_executors/condition.py and switch.py
simpleeval>=0.9.13

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Docker (MCP container management)
aiodocker>=0.22.0
